only copy meshes if they exist

+++ build.sh
@@ -51,9 +51,12 @@ touch "$DATA/${MODULE_NAME}.esp"
 
 # Copy back the module file and meshes into our directory to make sure we've got the latest source
 cp "$DATA/${MODULE_NAME}.esp" .
-MESHES="Meshes/${MODULE_NAME}"
-mkdir -p "$MESHES"
-cp -r "$DATA/$MESHES/"*.nif "$MESHES"
+if [[ -e "$DATA/$MESHES" ]]
+then
+  MESHES="Meshes/${MODULE_NAME}"
+  mkdir -p "$MESHES"
+  cp -r "$DATA/$MESHES/"*.nif "$MESHES"
+fi
 
 # Remove the temporary build folder
 rm -r "$DATA/$BUILD"
